Factor test strategies out into their own module.
